String to an integer

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -11,7 +11,7 @@
 int _atoi(char *s)
 {
 	int sign = 1, found_num = 0;
-	long int num = 0;
+	int num = 0;
 
 	while (*s != '\0')
 	{
